collector/requirements.txt
python_dateutil==2.8.2
pytz==2023.3
Requests==2.28.2
paho_mqtt==1.6.1
